Make PR template more suitable as a commit message (#549)
Currently, the PR template is very informative as a PR description, but
add some boiler plate to our commit messages especially if authors don't delete unused sections. This PR makes some very small changes to the template which make the PR description more suitable as a commit message.

Changes:
- Make the description be in sentence form by removing the
"Description:" boiler plate. I think it's pretty clear the prose at the top of a commit is a description.
-  Indicate that unused sections should be deleted
- Add a line for listing related github issues (we've been forgetting
to do this).