Allow runs beyond verification period in prognostic run diags (#536)
Currently the prog run diagnostic report code assumes (reasonably) that the prog run ends before the end of the 40-day SHiELD run, for the purposes of computing metrics. However we may want to extend prognostic runs beyond that time period for climate evaluation. This PR allows that to happen

Significant internal changes: Adds an inner join function that subsets both the prognostic run and the verification data to the common indices of both.