Parametrize cpu and memory requests for nudge-to-obs workflow (#560)
Want to do year-long nudge-to-obs fv3gfs run. Using only 6 cores is too slow. This PR parametrizes the cpu and memory requests for the relevant argo workflow to allow using more compute.

Added public API:

nudge-to-obs workflow template now takes optional parameters fv3gfs-cpu and fv3gfs-memory which control the cpu and memory requests for the fv3gfs part of the workflow. The defaults are the same as the current requests.