Save testing config, add metadata to ML report (#573)
It's difficult to know from just looking at the ML offline report what model is being tested and on what data it is being tested. This PR adds a metadata section to the report which includes the test config yaml (plus the model path). The timesteps used for testing are shown via two histograms.

Also, because of some defaults set for plotting (vmax and xlim), it was difficult to see any detail in many of the plots created by the report. The defaults are removed from the plotting routines in diagnostic_utils(but the limits can still be manually set from whatever function is calling these routines if desired).

Breaking changes:
- the offline_ml_diagnostics.create_report will fail if pointed to a diagnostics directory that was created from a previous fv3net commit (i.e. doesn't have config file saved)

Significant internal changes:
- The offline_ml_diagnostics.compute_diags script now saves a config.yaml in the output directory
- The offline_ml_diagnostics.create_report script now reads this config file and uses it to include metadata in the report
- Do not hard-code in default vmax or xlim for diagnostic_utils.plot_profile_var and diagnostic_utils.plot_column_integrated_var