Continue e2e test if job phase is Pending or null (#596)
The e2e test occasionally fails immediately. This is because the waitForComplete function in run_test.sh only allowed the workflow to proceed if its status is Running, but occasionally the status will be briefly Pending or null before it becomes Running. This PR makes waitForComplete keep watching the job if its status is Pending or null (but still times out in 30min).