Update prognostic and nudging runfiles for latest fv3gfs-wrapper image (#599)
The main purpose of this PR is to update the prognostic run infrastructure to use the latest fv3gfs-wrapper image.  This PR also updates some aspects associated with the nudge-to-fine workflow, namely:
- Updating the `nudging_chunks.yaml` file to include output tapes for `atmos_output`, `physics_output`, and `physics_tendency_output`.  
- Updating the nudging argo workflow to use the latest fv3gfs-wrapper image by default.
- Updating the `nudging_runfile.py` to include the eastward and northward winds in `STORE_NAMES`, which could help in learning momentum tendencies.
- Adding a fix for an issue in the volume definitions in the nudging argo workflow that prevented jobs from being submitted.

This PR does not touch anything associated with the nudge-to-obs workflow, which will need its runfile updated to be compatible with the latest python wrapper.