Centralize constraints for fv3net projects
Builds have been failing a lot lately because we are not comprehensively pinning our dependencies, and pinned versions are not being shared by the fv3fit, prognostic_run, and fv3net images. In some cases, these pinned versions conflict with each other. This PR concentrates all of the `pip` dependency version information into a `constraints.txt` file. `pip` can use this file to determine the versions of the packages it should download, but will not install all the packages in that file. This allows sharing pinned dependencies across a packages which do have the same requirements.

Significant internal changes:
- Move all version pinning information to `constraints.txt` and `environment.yml`
- Add linting check ensuring that no `requirements.txt` contains any version information.

Requirement changes:
- Fix pickling issue by pinning dask to 2.22.0
- Repair matplotlib issue by pinning to 3.2.2
- Pin docref to 0.2.7

- [ ] Open dask issue about pickling problem

Related #607