[VCMML-505] Add prognostic run diagnostics to combined template
Xarray cannot decode the metadata of the new segmented run outputs because the fill_value is set to 0 (#622). This problem would have been caught if the prognostic-run-diags were part of the end2end test, so this PR adds this check to the combined nudging-train-diag rule so that these outputs will be computed automatically for any prognostic run generated with this workflow. This may save some computation on the report generation side.

This PR also adds a failing unit test for the fill value issue, and fixes the problem.

- [x] Tests added

Resolves #622