Explicitly cache compiled fortran code when building prognostic run image (#635)
The new prognostic_run image takes a long time to build, owing in part to the time it takes to compile the fortran sources. While we hoped CircleCI docker layer caching would prevent re-building the intermediate layers, in practice they often need to be re-built with that system. In the longer term when the images are built on some kind of Google infrastructure we may have more reliable caching, but we need a way to shorten the build time in the short term.

This PR adds explicit caching of a compiled Fortran image when building the prognostic run image. This image is tagged based on the git sha in the fv3gfs-fortran submodule. When doing active development with potentially dirty files, you can force a re-build of the Fortran image using `make rebuild_image_fortran`.

Added public API:
- `make rebuild_image_fortran` forces a re-build of the compiled Fortran sources which are copied in to the prognostic run image

Significant internal changes:
- The prognostic run image now relies on an external compiled fortran image to produce the compiled sources it uses. These are compiled as part of the `make build_image_prognostic_run` target.

Minor changes:
- Updated references of fv3gfs-util to latest code for the fv3gfs_dockerfile refactor
- Updated fv3gfs-fortran to latest master

No tests are added, as this is a refactor.