Persist FV3 logs to disk
Save logs from the prognostic runs to disk. Previously, this information was only echoed to the console, but not persisted to disk. The `capture_output` flag to fv3run will save the output to file, but will not print to console. This PR does both using `tee`. This new functionality is tested by adding a new script `runfv3.sh` and ensuring it is used in both the prognostic run regression tests and the argo workflows.

- [x] Tests added

Resolves #624