Output data at constant interval from initial condition
The fortran and python diagnostics managers have different conventions for constant frequency output. The fortran model outputs every time at a the specified interval from the initial condition, but the python diagnostics are instead outputted at intervals from 0 UTC of the same day. This PR changes the python behavior to behave like the Fortran model, so its possible to sync up the Fortran and Python diagnostics.

This is technically a breaking change, but only for output frequencies which are not a factor of 1 day.

I also updated all the type hints to reflect that `fv3gfs.get_state(['time'])` is now a `cftime.DatetimeJulian` object. It is strange that this didn't give a type error before. @spencerkclark Is it possible that mypy interprets `cftime.DatetimeJulian` as an `Any` object?

- [x] Tests added