Use artifacts, not parameters for fv3config and runfile for run-fv3gfs (#632)
Argo has limits on the size (# of chars) in parameters which can be exceeded by large fv3config objects (e.g. a year-long nudge-to-obs run). This PR makes internal changes to the relevant argo workflows to use artifacts instead of parameters, which are not subject to this limit and are a more natural representation of these objects. Note that the artifact store is configured in `vcm-workflow-control/cluster` and these artifacts will be stored under `gs://vcm-ml-argo`.

No public API changes.

Significant internal changes:
- use artifacts instead of parameters
- delete "arguments" section of run-fv3gfs workflow template, since that template is not meant to be used as an entrypoint