Offline diags model path handling for train-diags-prog
This PR changes `offline_ml_diags` to be able to look for saved model data in the same directory as specified in as the output location of the train model workflow step, regardless of model training routine. This allows the `train-diags-prog` workflow to run without the hardcode of 'sklearn_model.pkl', so that fv3fit.keras may use this workflow (I missed this in #633 and the keras side of the workflow isn't currently tested). 

Refactored public API:
- Provide a default sklearn pkl filename to the load_sklearn_model function in the offline_ml_diags module and use the directory containing this file as the input argument instead. 
- Provide a default "model_data_dir" to the load_keras_model function in the offline_ml_diags module and use the directory containing this directory as the input argument instead. 
- Removes the somewhat redundant `model_loader` config argument from the training/offline config, which had been used to specify the model routine for the offline diags but had been made unusable by #591, and instead infers the appropriate model loader function from the `model_type` config argument.

- [X] Tests added