[VCMML-529] Serialize sklearn (#665)
We currently save the `SklearnWrapper` object using `joblib.dump`, which means that refactors to `fv3fit` code can easily break our ability to open the pickled object. This PR modifies the serialization format for scikit-learn models so that it only uses pickle for scikit-learn or builtin code, which should allow future versions of fv3fit to open this data more easily. The data are serialized to a single yaml file containing the serialized bytes of the scaler, the underlying sklearn models, and some metadata.

I also simplified the model loading infrastructure a little, and deleted some tests that were tightly scoped to implementation details of helper functions. I tried to write new tests covering the deleted cases, but instead targeting the public API of `SklearnWrapper`.

Added public API:
- `SklearnWrapper.dump/load`

- [x] Tests added