Count sea ice in ocean domain average
The land sea mask variable can change over time due to seasonal variations in sea ice. This PR lumps sea ice and sea surface types into the "sea" domain when taking domain averages in diagnostics, so grid cells will always be in the same domain averaging group over different seasons. 

It also fixes a related issue where the offline diags would fail for datasets with time-varying `land_sea_mask` when trying to merge the catalog grid `land_sea_mask` data (which was constant). The solution is to drop this variable from the grid before saving the output, since it is not used in the report- the reason the grid is merged before saving is so the lat/lon variables are available for map plotting.

Significant internal changes:
- Reversed key/value of the surface type enumeration in `diagnostics_utils.config` so that multiple floats can map to the same surface type "sea",  and made the appropriate change to the `diagnostics_utils.utils.snap_mask_to_type` which uses this mapping. This function is only used for the land sea mask mapping, so this doesn't break other parts of the code.


- [x] Tests added