Fix append_run chunk size validation (check target zarr, not source) (#671)
Currently the append_run script raises a value error if the length of the time dimension is not a multiple of the time chunk size for the source Zarr-store (i.e. the one that is being appended). However, for crashed segments, we cannot guarantee that the length will be a multiple of the chunk size. So this PR modifies the chunk-size validation to check the target Zarr-store and only raises a warning if the source store does not meet this criteria.

Significant internal changes:
- Update chunk size validation so that crashed run-fv3gfs segments can still be appended