Fix bug in offline diags metrics
The offline metrics calculation was broken because the area averaging function was altering the underlying data. This PR fixes the bug and adds more tests for the metrics functions. It also adds an optional arg to the offline report to save the commit hash it was generated with.

Added public API:
- `--commit-sha` optional arg in `offline_ml_diags.create_report`

Significant internal changes:
- Fix bug in area averaging
- Add test that will fail if underlying data is altered, make test cases that involve averaging use non-uniform weights


- [x] Tests added

Resolves #<github issues> [VCMML-552]


[VCMML-552]: https://vulcan.atlassian.net/browse/VCMML-552