Correct the 2020-10-02-getters-for-DeepConv experiment
This previous run  was run with a version of the fortran model that zero'd out
the tendencies storing them in the gfs physics diagnostics. This PR uses
a correct version of the model.

It also only runs for a single 10 day segment

argo: baseline-run-plx8b
outputs: gs://vcm-ml-archive/prognostic_runs/2020-10-02-getters-for-DeepConv