Minimal example of NaN failure
The NaN issue only occurs when using a loop object which is an instance
of a subclass of TimeLoop, AND which has a __init__ method.

For instance,  if "loop" is an instance of completely empty child class:

    class Subclass(TimeLoop):
        def __init__(self):
            TimeLoop.__init__(self)

The NaN errors will happen after the first dynamics time step.

OTOH, if "loop" is an instance of

    class Subclass(TimeLoop):
        pass

Then no such error occurs. This is incredibly bizarre.