[VCMML-566] Fix 24-processor prognostic run error
The prognostic run previously crashed on the first dynamics step
because of a very mysterious error message involving NaNs.

This error occurs for very minimal examples (see #686), only when
the wrapper is initialized after a large python import e.g. of runtime
or fv3fit.

This commit fixes this, but it's not clear why. It might be senstive to
the amount of memory used by python imports, and could be specific to
open mpi.