Add explicit encoding of chunk sizes in post-processing script
This is necessary to allow crashed segments to be appended properly to diagnostic zarr stores. Without this, the time-dimension for the crashed segment will have a chunk size equal to the length of the time dimension, and the resulting merged zarr-store will be invalid. The same issue can also occur for data variables depending on when the run crashes and the specified chunking.

Significant internal changes:
- Allow defining chunk sizes at runtime for nudging workflow. Necessary to allow short (1-hour) segmented run in e2e test.
- Explicitly set the chunks encoding of each variable within post-processing script
- Fix test_append.py (there was a bug in test_append_zarr_along_time, and need to add the explicit encoding here to make the test pass)

- [x] Tests added

Resolves #<github issues> [JIRA-TAG]

(Delete this for the commit message)
You are encouraged to check the PR against the [Code Review Checklist](https://paper.dropbox.com/doc/Code-Review-Checklist--A4lKrs~xg7w5Gsb39N6JLNQoAg-IlsYffZgTwyKEylty7NhY).