Add minimum amount to fv3fit scaler standard deviation
The standard scaler in fv3fit needs to be able to handle constant variables (inputs and outputs) without blowing up. It also needs to produce "reasonable" standard deviations for constant variables because std. dev is used in the default keras MSE loss function to normalize each level's contribution to the loss. Previously constants were assigned a std. dev. of 1.0, but this resulted in these levels being effectively ignored in NN training because the value is much larger than typical data's std. dev.  This PR instead adds a small value (default 10^-12) to the std. dev. of each variable in the scaler before normalizing/denormalizing. 

Along with the above, a small type conversion is included in the offline diags; both fixes are needed to get the nudge-to-obs dataset to work with fv3fit.keras training routine.

Refactored public API:
- fv3fit standard scaler now takes an input argument "std_epsilon" instead of "std_threshold"

Significant internal changes:
- std_epsilon now added to all variables' std dev. in standard scaler

- [X] Tests added