Disable docker-layer-caching and clean-up prognostic_run
Docker layer caching (DLC) is very expensive, and not well suited to the high parallelism of our build process. This PR changes the caching mechanism to use GCR (using the --cache-from flag). This new flag is only compatible for docker build kit builds, so I have modified the builds to use that instead. This will likely increase GCP network egress charges, as CircleCI will likely be pulling docker layers from GCR. This cost has not been significant in the past, but we should monitor any significant changes to our billing.

Unfortunately, the --cache-from flag must be set manually. When set correctly, the build time is around 5 minutes for the prognostic run image which is about the same performance as before. Once this image pushes to master, we should get good performance using "latest" as the tag.

Network charges would not occur if we migrate to google cloud builds. Since google cloud build seems to prefer one dockerfile builds, I have significantly streamlined the prognostic_run dockerfile by combining image stages, deleting the NCEPLibs build and Dockerfile.fortran, (which is in the FMS build image already), and basing the image off `us.gcr.io/vcm-ml/fms-build` rather than the ubuntu image. This change also ensures we are using a consistent version of the ubuntu OS to fms-build.