Improve performance of scikit learn serialization
Cramming all the data into a single yaml resulted in very slow read
performance. The yaml format is not really suitable for file sizes
larger than a MB, but the models are around 160MB. It is much faster to
save as separate files. This is a backwards incompatible change to the
saved model format, but necessary for interactively loading and
evaluating the ML models. This also makes the interface of `dumps` more
similar to the keras models (See #674).

Resolves #718