Exit from run-fv3gfs if output-url already exists (#722)
Currently if run-fv3gfs is set to output to a url that already has some content (e.g. a zarr store) it will try to append to the existing contents. This is confusing and unexpected behavior. This PR adds a check before the first step of the run-fv3gfs workflow that the output url is empty and exits with an error message if not.