Allow nudging towards a coarser dataset
Previously, the nudging runfile assumed that every time is present in
the nuding dataset. This severly constrains the model timestep, and in
particular, the nudging runfile will not work when the FV3 timestep is
less than that of the nudging reference dataset.

This approach is also innefficient, and does much more I/O than is
probably necessary given the smooth nature of the nudging tendency.

This PR adds options to the nudging configuration to specify the
frequency at which target data will be retrieved. This can also be used
to subsample a target dataset with full output frequency.