Don't mount gfs-analysis PVC, bump CI argo to v2.11.6 (#735)
Mounting this persistent volume claim is unreliable. We may want to move away from mounting PVC's at all, but for now, this PR switches all workflows to mount the nudging-read-only PVC (which contains the coarsened restarts) since that volume is for some reason more reliable. The downside of this change is that runs which are nudged towards GFS analysis will have to download all the required files from GCS before the run starts.