Refactor FV3DerivedState and add vcm.DerivedMapping
Previously, the prognostic run used its own `runtime.DerivedFV3State` object for convenient access to derived variables, e.g. cos(z). Since this would also be useful for other cases where derived variables need to be used, i.e. training and diagnostics, this PR introduces a class `DerivedMapping` from which the prognostic run `DerivedFV3State` is composed. A new `FV3StateMapper` is introduced which wraps the getter so that `DerivedFV3State` can make use of the `__getitem__` interface. Derived variables specific to the runtime usage are defined in the `FV3StateMapper`.

This PR is just to add `DerivedMapping` to vcm and refactor `DerivedFV3State` to use it and the `FV3StateMapper`. There is more changes that will occur in the the training and diagnostic workflows to make use of the `DerivedMapping`, which will be added in a subsequent PR. Nothing has changed about the `DerivedFV3State` functionality, other than location and that it does not register functions anymore. 

Added public API:
- Created new `DerivedMapping` in `vcm`

Significant internal changes:
- Refactor of `DerivedFV3State` to use `DerivedMapping` and `FV3StateMapper`
- Added the `FV3StateMapper` class. 
- small change to the registered `cos_zenith_angle` derived variable function so that it can take in a data array of times

Resolves #<github issues> [VCMML-620]



[VCMML-620]: https://vulcan.atlassian.net/browse/VCMML-620