Allow custom spec of python output frequency in nudge-to-obs
It can be useful to be able to specify different frequencies for outputs that are used for ML versus outputs that are used to diagnose the climate of a training run. E.g. for ML it's nice to use a frequency that does not evenly divide 24 hours so that the resultant samples will span the diurnal cycle. But for other diagnostics which will ultimately be used for prognostic verification purposes, it is much more convenient to use a more conventional 3hr or 6hr output frequency.

Adds this option for nudge-2-obs in a backwards compatible way.