Package the intake catalog yml in vcm (#771)
The intake catalog is used a lot of places, but there is no consistent or convenient way to specify it's path. For some workflows, any script must be run from the fv3net root. In addition, loading data from the catalog outside of the fv3net context (e.g. in explore) requires hard-coding a local path. There is also a lot of duplicated code for specifying it's path.

This PR packages the catalog.yml file inside of vcm so that it can easily be referred to wherever vcm is installed. I chose to add it to vcm since vcm is imported everywhere the catalog is used, and the vcm dependencies are similar to the catalog's dependencies.

Some defunct code in vcm.convenience and vcm.regrid was deleted as part of the refactor.

Added public API:

    vcm.catalog.catalog: the catalog
    vcm.catalog.catalog_path: the path to the catalog