Compute exposed area and add metadata (#784)
This PR adds an "exposed_area" variable to the coarsened output representing the area of a grid cell that lies above the surface. This is needed to formulate heat and moisture budgets accurately over topography. This [notebook](https://gist.github.com/nbren12/cd2322d3913a8bb1f246d72884bdb31e) contains an example plot of this diagnostic.

This code assumes that the "exposed area" at C384 is constant in time on model levels. This allows using a single time-step dataset produced by @spencerkclark to compute exposed areas for all times. However, any additional coarse-grained area loss from C384-C48 is time dependent.

Unfortunately, this required adding yet another data-set to be merged as part of this workflow, and more bespoke merging logic. That said, this workflow almost always needs some reworking for a new SHiELD run, so this problem is not new.