Add movies and spatial-mean diagnostics of vertical mean dQu/dQv (#789)
Useful to be able to have some idea of what online predictions of dQu and dQv are in our prognostic runs. Now if the make-movies option is set to true for the prognostic run report, a movie vertical mean dQu and dQv will be made. We don't have an easy way to diagnose pQu or pQv right now, so the movie just shows the ML contributions.

Significant internal changes:
- make movie of vertical mean dQu and dQv
- add dQu and dQv to diagnostic plots of spatial/zonal means