Add a MultiDatasetMapper for training using multiple datasets (#774)
This PR adds a `MultiDatasetMapper` which can be used for training an ML model on multiple datasets.  Currently it requires that the datasets have timesteps in common with each other.

Added public API:
- `loaders.mappers.open_merged_nudged_full_tendencies_multiple_datasets`, a function that takes in a sequence of paths to nudged-to-fine training datasets and returns a `MultiDatasetMapper`.

Significant internal changes:
- `loaders._utils.stack_dropnan_shuffle` now can handle input datasets with a `"dataset"` dimension.  Currently it is set to automatically thin the resulting batch by a factor of the length of the `"dataset"` dimension, if one exists, such that the number of columns used in training remains constant per number of timesteps included, whether one is training with a single dataset or multiple.
- Within the offline diagnostics, if a `"dataset"` dimension exists, we now stack it with the `"time"` dimension and refer to the combined dimension as the `"sample"` dimension.
- `predict_columnwise` allows the `"dataset"` dimension to be broadcast in the stacking process used to generate samples.
- Added optional `memory` parameters for the training and offline-diags workflows, and exposed them as optional parameters in the train-diags-prog workflow as `memory-training` and `memory-offline-diags`, respectively.

This was tested in a `train-diags-prog` argo workflow and everything ran to completion; see the configs in explore [here](https://github.com/VulcanClimateModeling/explore/tree/master/spencerc/2020-11-11-multi-dataset-mapper-yml).  Note that it required increasing the memory allocated in the training and offline-diags workflows to 12 GB and 20 GB, respectively (I suspect more memory would be required if more than three datasets were used).