By default use first test timestep as snapshot time (#830)
Currently the offline diags uses the first key in the sorted pred_mapper keys for the snapshot transect plot. However this key can fall outside the user's specified test timesteps. This PR makes the snapshot time be the first item in the sorted timesteps list (which will be all the keys in pred_mapper if a custom set of timesteps is not provided).