New nudge-to-x mappers (#821)
With nudge-to-fine runs now (mostly) capable of being conducted via the sklearn_runfile and its output data simplified, new simpler mappers are needed to the resulting datasets. This PR adds those mappers as well as new mappers for the runfile's nudge-to-obs option. It also deprecates but doesn't remove the legacy nudged dataset mappers. It adds `synth`-based unit tests for new mappers, while maintaining the existing, more extensive test suite on the legacy mappers. While this PR does not touch the runfile or the argo worflows (this is left to a subsequent PR), the new mappers do establish data specs for the runfile's output for nudge-to-fine and nudge-to-obs, which are detailed in `workflows/prognostic_c48_run/OUTPUTDATA.md`.

Added public API:
- `loaders.mappers` now has `open_nudge_to_fine`, `open_nudge_to_fine_multiple_datasets`, and `open_nudge_to_obs` mapper functions as the means of accessing data from the nudge-to-x sklearn_runfile. These use `XarrayMapper` for a more intuitive computational process.

Refactored public API:
- Previous opener functions from `loader.mappers._nudged` have been refactored into a `loaders.mappers._nudged._legacy` module, and while nothing has been removed from the `loader.mappers` API, these functions are now considered deprecated and generate a warning.
- The `SubsetTimes` object is refactored to `loaders.mappers._transformations`.
- The unused `GroupByTime` object in the legacy nudging mapper module is deleted, as it was duplicated the functionality of the group by time object in `_fine_resolution_budget`.

- [X] Tests added

Resolves [VCMML-664]