Harmonize .fit methods of all fv3fit models
To address Brian's comments about adding a dump method to "Predictor",
we decided to base the SklearnWrapper off of fv3fit.keras.Model, which
is a generic base class, that for some reason was hiding in the keras
namespace.

Unfortunately, this required addressing an inconsistency in the .fit
methods of the scikit learn and keras models.  SklearnWrapper.fit takes
Dataset, but KerasModel.fit takes Sequence[Dataset].

Because initializing the scaler requires reading some data, I also moved
it's initailized into SklearnWrapper.fit.