N2F in sklearn_runfile in e2e, delete nudging runfile (#836)
This PR moves the nudge-to-fine (N2F) step of the e2e test to the sklearn_runfile, and deletes the nudging runfile and argo workflow that deployed it. It allows mass changes in the delp nudging case only, and establishes a way of configuring default diagnostic outputs for all of the runfile cases. It also adds output of model state and reference state variables to the sklearn_runfile in N2F runs, which had been missing from #791. As an example of the output of a N2F run in the e2e test, see `gs://vcm-ml-scratch/test-end-to-end-integration/integration-test-881d911ba233/nudge_to_fine_run`, following data specs in `workflows/prognostic_c48_run/OUTPUTDATA.md`.

Added public API:
- Enhances the runtime's `DiagnosticFile` object to be useful in configuring default data formats (e.g., for how physics and nudging tendencies, diagnostics, and model states are written out to zarrs) for the different run types.

Refactored public API:
- Deletes workflows/prognostic_c48_runs/nudging/nudging_runfile.py and associated files
- Deletes workflows/argo/nudging/nudging.yaml

Significant internal changes:
- e2e test uses the `prognostic-run` workflow and the sklearn_runfile for the N2F section and `open_nudge_to_fine` for model training/offline diags
- `runtime.FV3StateMapper` allows mass-nonconserving updates for the pressure thickness variable
- `runtime.DiagnosticFiles` used to store default diagnostic zarr names and variables for preparing config

- [X] Tests added

Resolves  [VCMML-694]