Plot jacobian in report
Adds jacobian plots to the report. Example: http://storage.googleapis.com/vcm-ml-public/noah/2020-12-16/index.html

I renamed the html file to index.html.

I chose to save pngs of the jacobian in the `compute_diags` step, and then have the `create_report` copy any existing pngs in the "diagnostics" folder into the report as is under the heading "Other Figures". This might be a good pattern for things like the transect plots in the future where we don't actually need access to the quantitative information.

Added public API:
- `fv3fit.PackedKerasModel.jacobian` computes the jacobian

- [x] Tests added