Time averaged outputs (let's try that again)
#898 broke the integration tests, and this PR fixes that, and adds a faster check to the prognostic run tests that finds the error outside of the end-to-end. Most of these changes were reviewed already. This is the relevant diff: f8e87d3...3aa977d.

Because the DiagnosticFiles needed to accumulate data that is averaged, they need to figure out when an interval is complete. Normally, this occurs when a timestep from the next interval is observed, which doesn't happen for the final interval. In that case, one needs to explicitly flush the data. 

I put this in the `__del__` of `DiagnosticFile`, but apparently calling `fv3gfs.wrapper.cleanup()` short-circuits the standard python cleanup. That could cause other problems.

Other changes:
- Remove "time" from the "state_after_timestep" output of the nudging run since this raises an error when averaging.