Option to save epoch checkpoints
This adds the option to checkpoint keras models at each training epoch. To do so, set the field `save_model_checkpoints` to `True` in the training config file. If this field is not provided the default behavior is to not save checkpoints. Checkpoints will be saved to the subdir `model_checkpoints` in the model training directory.

Minor change to the offline diags argo template to fix an indentation error that was preventing the memory parameter from getting used.

Added public API:
- optional `save_model_checkpoints` field in training config. Only applicable for keras training.