Fix: remove gcs access (#927)
Fixes an issue that broke master when I merged the 'remove GCS access from fv3fit tests' PR earlier.

The problem wasn't apparent when I tested locally because I used a different config than the integration test, which did not include a derived variable in the test feature list. The issue was that the offline diags PredictionMapper creates a DerivedMapping in order to have derived inputs available for predictions, and then the timestep values from PredictionMapper are used as the datasets in the batch loader, which checks if the variables that rely on the grid exist in the dataset and only loads the grid if they are absent from data_vars. However, if the dataset was produced by a DerivedMapping, it may have derived variables such as cos_zenith_angle in its data_vars but be missing the variables ("lon" in that example) that are needed to derive it.

To account for this, the grid will be loaded if any derived variables that need it are present in the requested_variables list, even if they are in the data_vars.