Add .diagnostics[].times.includes_lower option
The model currently includes the starting time of an interval in the
average, but not the ending time. Because the diagnostics are passed the
time at the end of a timestep. This means that a run of n averaging
intervals will currently save n + 1 outputs. n completed intervals and 1
"average" including only the values at the final timestep.

This option allows one to control if the interval over which an average
is computed includes it starting or ending time. This allows a run of n
averaging intervals to save n outputs, and changes this to the default
behavior.