Fix nudging config interp (#939)
The nudge-to-fine time interpolation routine did not work in the sklearn_runfile. This was because the interpolation wrapper added `time` to the state dict unlike the non-interpolated states, and because the interpolation couldn't be configured after #915. This PR resolves both of those issues and allows time-interpolated nudging runs. However, note that this PR doesn't yet resolve the zarr monitor writing issue (fv3net nine-three-eight) but does get master to a point where that issue is easily reproducible so that we can fix it.

Refactored public API:
- the `_time_interpolate_func` wrapper no longer adds its input argument `time` to the returned dict, since it's by definition already available to the user