Fill missing vars with NaNs for prognostic report
On the prognostic report, if the HoloViews plot are set to display info for a variable which does not exist for all of the grouped runs, the plot often shows nonsensical data for the runs which do not have that diagnostic computed (often the verification). This PR resolves that issue by inserting NaN-filled arrays for any "missing" variables. [Here](https://storage.googleapis.com/vcm-ml-public/argo/2021-01-24-n2o-c48-c96-prognostic-comparison/redo-report/index.html) is an example report generated with this branch.

Compare an example before:
![Screen Shot 2021-01-28 at 2 40 10 PM](https://user-images.githubusercontent.com/23640870/106207911-5e897000-6177-11eb-830e-5564e8e1f6b4.png)

and after:
![Screen Shot 2021-01-28 at 2 40 01 PM](https://user-images.githubusercontent.com/23640870/106207888-54677180-6177-11eb-986e-79a8552bbe4e.png)

Significant internal changes:
- add function to insert NaN-filled arrays as necessary

- [x] Tests added